+++ R/LoopThruSpectra.R
@@ -1,5 +1,5 @@
 #'
-#' Display the Spectra in a Spectra Object One at a Time
+#' Display the Spectra in a Spectra Object One at a Time DEPRECATED
 #'
 #' Plots each spectrum in a \code{\link{Spectra}} object one at a time, and
 #' waits for a return in the console before plotting the next spectrum.  Use
@@ -39,40 +39,5 @@
 #' }
 #'
 loopThruSpectra <- function(spectra, ...) {
-  .chkArgs(mode = 11L)
-
-  go <- chkGraphicsOpt()
-  if (go == "base") {
-    cat("Press ESC to stop looping through the spectra\n\n")
-    ns <- length(spectra$names)
-    for (i in 1:ns) {
-      tt <- paste(spectra$names[i], " (#", i, " of ", ns, ")", sep = "")
-      plotSpectra(spectra, which = i, main = tt, ...)
-      devAskNewPage(ask = TRUE)
-    }
-    devAskNewPage(ask = FALSE)
-  }
-
-  if (go == "ggplot2") {
-    x <- spectra$freq
-    l.x <- length(x)
-    df1 <- data.frame(x = NA_real_, y = NA_real_, spectra.name = NA_real_)
-    for (i in 1:length(spectra$names))
-    {
-      y <- spectra$data[i, ]
-      spectra.name <- rep(spectra$names[i], l.x)
-      df2 <- data.frame(x = x, y = y, spectra.name = spectra.name)
-      df1 <- rbind(df1, df2)
-    }
-    df1 <- df1[-1, ]
-    p <- ggplot(df1, aes(x = x)) +
-      geom_line(aes(y = y)) +
-      theme_bw() +
-      xlab(spectra$unit[1]) +
-      ylab(spectra$unit[2]) +
-      facet_grid(spectra.name ~ ., switch = "both") + # faceting
-      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())
-
-    return(p)
-  }
+  stop("loopThruSpectra has been renamed reviewAllSpectra")
 }