+++ NEWS.md
@@ -2,6 +2,7 @@
 ## Misc.
 * Courtesy GSOC and Tejasvi Gupta, the following plotting functions gain the ability to produce either `base` graphics or `ggplot2` graphics: `plotSpectra`, `surveySpectra`, `loopThruSpectra` (which has been renamed to `reviewAllSpectra`).
 * The graphics output options can be chosen by `options(ChemoSpecGraphics = "option")`; details in new help file `?GraphicsOptions`.
+* `loopThruSpectra` has been renamed `reviewAllSpectra` to better reflect what it does under the new graphics options.
 
 # ChemoSpec 5.3.11 2021-03-24
 ## Changes in ChemoSpecUtils that affect ChemoSpec